v 0.9.4
* Test before emptying cache if in a Magento root dir

+++ inc/empty-cache.sh
@@ -4,12 +4,19 @@
 ## Cache
 ###################################
 
+if [ ! -f "api.php" ]; then
+    echo "/!\ You are not in a Magento root dir /!\\";
+fi;
 
-echo "- Clear cache";
-cp "${SOURCEDIR}/files/magetools_flush_cache.php" "shell/magetools_flush_cache.php";
-php -f "shell/magetools_flush_cache.php";
-rm -f "shell/magetools_flush_cache.php";
+if [[ -d "shell/" ]]; then
+    echo "- Clear cache";
+    cp "${SOURCEDIR}/files/magetools_flush_cache.php" "shell/magetools_flush_cache.php";
+    php -f "shell/magetools_flush_cache.php";
+    rm -f "shell/magetools_flush_cache.php";
+fi;
 
-echo "- Deleting var/cache";
-rm -rf "var/cache/";
-rm -rf "var/full_page_cache/";
+if [[ -d "var/" ]]; then
+    echo "- Deleting var/cache";
+    rm -rf "var/cache/";
+    rm -rf "var/full_page_cache/";
+fi;
